Move generic dataset outside of torchtext
There is no reason for it to live inside torchtext, and having it
outside will help with using torchtext as a library
