+++ decanlp/text/torchtext/datasets/__init__.py
@@ -5,8 +5,6 @@ from .translation import TranslationDataset, Multi30k, IWSLT, WMT14  # NOQA
 from .sequence_tagging import SequenceTaggingDataset, UDPOS # NOQA
 from .trec import TREC
 from .imdb import IMDb
-from . import generic
-
 
 __all__ = ['LanguageModelingDataset',
            'SNLI',
